Added a fix if routing comes from ilrepositorygui instead of ilobjectplugingui as base class
This happens, when a link for a xcwi-object is created in the breadcrumbs (or maybe elsewhere as well). In this case, the method ->printToStdOut is called in ilobjectplugingui and ilrepositorygui. The page is therefore printed twice to std out and makes the ILIAS Page a little bit buggy (form-buttons stop working etc.)